rename Tooltip2 to StructuredPopover
