+++ datahub-web-react/src/alchemy-components/components/StructuredPopover/index.ts
@@ -1,0 +1,1 @@
+export { StructuredPopover } from './StructuredPopover';
