--- datahub-web-react/src/alchemy-components/components/Tooltip2/index.ts
@@ -1,1 +1,0 @@
-export { Tooltip2 } from './Tooltip2';
