+++ datahub-web-react/src/alchemy-components/index.ts
@@ -38,4 +38,4 @@ export * from './components/Tooltip';
 export * from './components/Utils';
 export * from './components/WhiskerChart';
 export * from './components/ColorPicker';
-export * from './components/Tooltip2';
+export * from './components/StructuredPopover';
