+++ datahub-web-react/src/app/entityV2/shared/versioning/VersionsDrawer.tsx
@@ -8,7 +8,7 @@ import { Link } from 'react-router-dom';
 import { useDebounce } from 'react-use';
 import styled from 'styled-components';
 
-import { Tooltip2 } from '@components/components/Tooltip2';
+import { StructuredPopover } from '@components/components/StructuredPopover';
 
 import { useEntityContext } from '@app/entity/shared/EntityContext';
 import { VersionPill } from '@app/entityV2/shared/versioning/common';
@@ -163,7 +163,7 @@ export default function VersionsDrawer({ versionSetUrn, open }: Props) {
             ),
             comment: <Typography.Text ellipsis={{ tooltip: true }}>{versionProperties?.comment}</Typography.Text>,
             createdAt: (
-                <Tooltip2
+                <StructuredPopover
                     width={250}
                     placement="topRight"
                     sections={[
@@ -178,7 +178,7 @@ export default function VersionsDrawer({ versionSetUrn, open }: Props) {
                     ]}
                 >
                     {moment(versionProperties?.created?.time).fromNow()}
-                </Tooltip2>
+                </StructuredPopover>
             ),
             menu: (
                 <StyledDropdown
